requirements.txt
aiohttp
spotipy
yt-dlp[default]
requests
aiogram
python-dotenv
pillow
pytest